feat: validate uid
feat: validate required fields

fix: change email to hobbies

+++ src/routes/users.ts
@@ -1,5 +1,6 @@
 import type { IncomingMessage, ServerResponse } from "http";
 import { db } from "../db/database.ts";
+import { validate } from "uuid";
 
 export const handleUserRequest = async (req: IncomingMessage, res: ServerResponse) => {
     res.setHeader("Content-Type", "application/json");
@@ -12,6 +13,10 @@ export const handleUserRequest = async (req: IncomingMessage, res: ServerRespons
                 return res.end(JSON.stringify(users));
             } else if (req.url?.startsWith("/api/users/")) {
                 const id = req.url.split("/")[3];
+                if (!validate(id)) {
+                    res.writeHead(400);
+                    return res.end(JSON.stringify({ error: "Invalid UUID format." }));
+                }
                 const user = await db.getById(id);
                 if (user) {
                     res.writeHead(200);
@@ -30,24 +35,36 @@ export const handleUserRequest = async (req: IncomingMessage, res: ServerRespons
                 });
                 req.on("end", async () => {
                     console.log("Received body:", body);
-                    const { name, age, email } = JSON.parse(body);
-                    const newUser = await db.create(name, age, email);
+                    const { username, age, hobbies } = JSON.parse(body);
+                    if (!username || !age || !hobbies || hobbies.length === 0) {
+                        res.writeHead(400);
+                        return res.end(JSON.stringify({ error: "Missing required fields" }));
+                    }
+                    const newUser = await db.create(username, age, hobbies);
                     res.writeHead(201);
                     return res.end(JSON.stringify(newUser));
                 });
             }
             break;
-        case "PUT":            
+        case "PUT":
             if (req.url?.startsWith("/api/users/")) {
                 const id = req.url.split("/")[3];
+                if (!validate(id)) {
+                    res.writeHead(400);
+                    return res.end(JSON.stringify({ error: "Invalid UUID format." }));
+                }
                 let body = "";
                 req.on("data", chunk => {
                     body += chunk.toString();
                 });
                 req.on("end", async () => {
                     console.log("Received body:", body);
-                    const { name, age, email } = JSON.parse(body);
-                    const updatedUser = await db.update(id, name, age, email);
+                    const { username, age, hobbies } = JSON.parse(body);
+                    if (!username || !age || !hobbies || hobbies.length === 0) {
+                        res.writeHead(400);
+                        return res.end(JSON.stringify({ error: "Missing required fields" }));
+                    }
+                    const updatedUser = await db.update(id, username, age, hobbies);
                     if (updatedUser) {
                         res.writeHead(200);
                         return res.end(JSON.stringify(updatedUser));
@@ -61,6 +78,10 @@ export const handleUserRequest = async (req: IncomingMessage, res: ServerRespons
         case "DELETE":
             if (req.url?.startsWith("/api/users/")) {
                 const id = req.url.split("/")[3];
+                if (!validate(id)) {
+                    res.writeHead(400);
+                    return res.end(JSON.stringify({ error: "Invalid UUID format." }));
+                }
                 const deleted = await db.delete(id);
                 if (deleted) {
                     res.writeHead(204);
